+++ repository_manager/backend/src/routes/commit.ts
@@ -41,7 +41,7 @@ commitRouter.get('/hash/:commitHash/pull', async (req, res) => {
                     include: {
                         collection: true
                     }
-                },
+                },              
             }
         });
         if (!commit) {
@@ -73,11 +73,29 @@ commitRouter.get('/hash/:commitHash', async (req, res) => {
     }
 });
 
-
 // pull the latest commit
 commitRouter.get('/latest', async (req, res) => {
     try {
-        const latestCommit = await prisma.commit.findFirst({ orderBy: { createdAt: 'desc' } });
+        const { branchId } = req;
+        // latest commit in this branch
+        const latestCommit = await prisma.commit.findFirst({
+            where: { branchId },
+            orderBy: { createdAt: 'desc', },
+            include: {
+                branch: true,
+                committer: true,
+                params: {
+                    include: {
+                        ZKMLProof: true
+                    }
+                },
+                nft: {
+                    include: {
+                        collection: true
+                    }
+                },              
+            }
+        });
         if (!latestCommit) {
             res.status(404).send({ error: { message: 'No commits found.' } });
             return;
@@ -132,6 +150,7 @@ commitRouter.post('/create', async (req, res) => {
             paramHash,
             params,
             metrics,
+            architecture,            // architecture of the model is a new requied field now
             commitHash,
             acceptedCommits,         // the hash array of the accepted commits if it is a merger commit
             rejectedCommits,         // the rejected commits with the hash and the reject message
@@ -142,15 +161,22 @@ commitRouter.post('/create', async (req, res) => {
                 paramHash: string,
                 params: commitParameters,
                 metrics: commitMetrics,
+                architecture: string,
                 commitHash: string,
                 // reserved for the system commits
                 acceptedCommits: string[] | undefined,          // accepted and rejected commits are undefined for general commits
                 rejectedCommits: RejectedCommits | undefined,   // only needed in case of system commits 
             } = req.body;
 
+
+
+
+        // --------------------------------- Commit Parameter validation section ----------------------------------------------------- //            
+
+        let warnings = [];              // in case a warning needs to be sent along with the response            
         // cannot create a commit if the base model is not uploaded
         const repo = await prisma.repository.findUnique({ where: { id: req.repoId } });
-        if (!repo!.baseModelHash) {
+        if (!repo!.baseModel || !repo!.baseModel) {
             res.status(400).send({ error: { message: 'Base model not uploaded. Cannot create commit.' } });
             return;
         }
@@ -162,8 +188,10 @@ commitRouter.post('/create', async (req, res) => {
         if (!branchId) {
             throw new Error('Criticial Error: branchId not attached to response.');
         }
+
         // Validate input fields
-        if (!paramHash || !params || !metrics || !commitHash) {
+        // in this version the metrics is an optional field
+        if (!paramHash || !params || !architecture || !commitHash) {
             res.status(400).send({ error: { message: 'Complete commit information not provided.' } });
             return;
         }
@@ -190,8 +218,7 @@ commitRouter.post('/create', async (req, res) => {
 
         const repository = branch.repository;
         // Validate write permissions
-        const hasWriteAccess =
-            repository.ownerAddress === pk || repository.writeAccessIds.includes(pk);
+        const hasWriteAccess = repository.ownerAddress === pk || repository.writeAccessIds.includes(pk);
 
         if (!hasWriteAccess) {
             res.status(403).send({
@@ -200,8 +227,75 @@ commitRouter.post('/create', async (req, res) => {
             return;
         }
 
-        // get the hash of the latest accepted commit and attach it as the previous hash of this commit
-        // for the first commit in the branch, this will return null        
+        if (!params.params) {
+            res.status(400).send({ error: { message: 'Error: No parameters (weights) provided.' } });
+            return;
+        }
+        const { zkmlProof } = params;
+        if (!zkmlProof) {
+            warnings.push('No ZKML proof provided for commit.');
+            // res.status(400).send({ error: { message: 'Error: No ZKML proof provided as params.' } });
+            // return;
+        }
+        else {
+            // ------------------- Uncomment if you want the ZKML field to be mandatory -------------------------
+            // // extract the zkml proof for this commit
+            const { verifierKey, circuitSettingsSer, proofSer, srsSer } = zkmlProof;
+            // // additional checks for the zkml proof settings
+            // if (!verifierKey) {
+            //     res.status(400).send({ error: { message: 'No ZKML verifier key found in the commit.' } });
+            //     return;
+            // }
+            // if (!circuitSettingsSer) {
+            //     res.status(400).send({ error: { message: 'No ZKML circuit settings Ser found in the commit.' } });
+            //     return;
+            // }
+            // if (!proofSer) {
+            //     res.status(400).send({ error: { message: "No ZKML proof Ser found in the commit." } });
+            //     return;
+            // }
+            // if (!srsSer) {
+            //     res.status(400).send({ error: { message: 'No ZKML srsSer found in the commit.' } });
+            //     return;
+            // }
+            // ZKML fields are not mandatory but if provided they must be unique
+            if (verifierKey && circuitSettingsSer && proofSer && srsSer) {
+                const sameZKMLProofs = await prisma.zKMLProof.count({
+                    where: {
+                        verifierKey,
+                        circuitSettingsSer,
+                        proofSer,
+                        srsSer
+                    }
+                });
+                // if same zkml proofs are found the commit cannot be created
+                if (sameZKMLProofs) {
+                    res.status(400).send({ error: { message: 'ZKML proof already exists. ZKML proof must be unique.' } });
+                    return;
+                }
+            }
+        }
+
+        // maintaing uniqueness of the commits
+        // finally we need to make sure that the paramHash commitHash and parameters are unique for each commit
+        const sameHashCommits = await prisma.commit.count({ where: { commitHash } });
+        if (sameHashCommits) {
+            res.status(400).send({ error: { message: 'Commit hash already exists. Commit hash must be unique.' } });
+            return;
+        }
+        const sameParamHashCommits = await prisma.commit.count({ where: { paramHash } });
+        if (sameParamHashCommits) {
+            res.status(400).send({ error: { message: 'Paremeter hash of commit already exists. Parameter hash must be unique.' } });
+            return;
+        }
+        const sameParams = await prisma.params.count({ where: { params: params.params } });
+        if (sameParams) {
+            res.status(400).send({ error: { message: 'Model parameters of commit already exists. Paramerers must be unique.' } });
+            return;
+        }
+
+
+        // get the parent merger commit for the commit to create
         const latestMergerCommit = await prisma.commit.findFirst({
             where: { status: 'MERGERCOMMIT', branchId },        // latest merger commit in this branch
             orderBy: { createdAt: 'desc' }
@@ -213,23 +307,23 @@ commitRouter.post('/create', async (req, res) => {
 
         // const previousMergerCommit = (latestMergerCommit !== null) ? latestMergerCommit.commitHash : (commitLength == 0) ? process.env.GENESIS_HASH : undefined;
         if (!previousMergerCommit) {
-            console.error('Unresolved conflict. Previous commit hash of non-empty branch is undefined.');
+            console.error('Critical Error: Previous commit hash of non-empty branch is undefined.');
             res.status(500).send({ error: { message: 'Internal Server Error' } });
             return;
         }
 
 
         // the first commit is the only commit that is committed by the user and still a MERGER_COMMIT
-
         // Commit message section
         // Default status for a new commit is pending unless it is the first commit in the branch for which it is the merger commit
         const commitLength = await prisma.commit.count({ where: { branchId } });
         const status = commitType == 'SYSTEM' ? 'MERGERCOMMIT' : commitLength == 0 ? 'MERGERCOMMIT' : 'PENDING';
         const commitMessage = commitType == 'SYSTEM' ? '_SYSTEM_COMMIT_' : message;
 
+        //  ---------------------------------- Handling of merger commits comes here --------------------------------------
 
-        // if it is a merger commit
         if (commitType == 'SYSTEM') {
+
             if (!acceptedCommits || !rejectedCommits) {
                 res.status(400).send({ error: { message: 'Accepted and Rejected Commits are mandatory for a merger commit.' } });
                 return;
@@ -257,15 +351,14 @@ commitRouter.post('/create', async (req, res) => {
                 });
 
                 // ------------------------------------ ZKML Proof handling --------------------------------------------
-
                 // !-- this deletes the verified proof since we have already verified it
                 // !!-- the ZKML proof is a big file to store. Currently we are removing it once the commits are accepted.
                 // !!------          But change this later on need            --------------
+
                 if (updatedCommit.params) {
 
                     // in reality all commits must contain ZKML proofs and its verification cannot be skipped                    
                     // the merger must check if all the ZKML proofs have been verified before creating a Merger Commit !
-
 
                     // !! -- For Testing we consider that the ZKML proof may be empty
                     // !! But in production environments this cannot be empty
@@ -275,11 +368,13 @@ commitRouter.post('/create', async (req, res) => {
                         await prisma.zKMLProof.delete({
                             where: { paramId: updatedCommit.params.id }
                         });
-                    else console.warn(`High Level Warning!! Commit that does not contain ZKML proof has been merged. Commit Details: 
+                    else {
+                        warnings.push(`High Level Warning!! Commit that does not contain ZKML proof has been merged. Commit Details: 
                         \n Commit Hash: ${updatedCommit.commitHash}
                         \n Committed by ID: ${updatedCommit.committerId}
                         \n Commit Message: ${updatedCommit.message}
-                        \n Commit Time: ${updatedCommit.createdAt}`)                        
+                        \n Commit Time: ${updatedCommit.createdAt}`);
+                    }
                 }
             }
 
@@ -297,23 +392,17 @@ commitRouter.post('/create', async (req, res) => {
 
             // for the rejected commits delete all the parameters for those commits
             for (const id of rejectedCommitIds) {
-                const deletedParam = await prisma.params.delete({ where: { commitId: id } });
-                
-                // also delete the associated zkml proof for this param if it exists 
-                
-                // For rejected commits it is allowable to have the ZKML proof as optional
-                const rejectedZKMLProofCount = await prisma.zKMLProof.count({where: {paramId: deletedParam.id}})
-                if (rejectedZKMLProofCount > 0)
-                    await prisma.zKMLProof.delete({ where: { paramId: deletedParam.id } });
+                const paramToDelete = await prisma.params.findUnique({ where: { commitId: id } });
+                if (paramToDelete) {
+                    await prisma.params.delete({ where: { commitId: id } });
+                    // also delete the associated zkml proof for this param if it exists 
+                    // For rejected commits it is allowable to have the ZKML proof as optional
+                    const rejectedZKMLProofCount = await prisma.zKMLProof.count({ where: { paramId: paramToDelete.id } })
+                    if (rejectedZKMLProofCount > 0)
+                        await prisma.zKMLProof.delete({ where: { paramId: paramToDelete.id } });
+                }
             }
         }
-
-        if (!params.zkmlProof) {
-            res.status(400).send({ error: { message: 'Error: No ZKML proof provided as params.' } });
-            return
-        }
-        // extract the zkml proof for this commit
-        const { verifierKey, circuitSettingsSer, proofSer, srsSer } = params.zkmlProof;
 
         // Finally create the commit
         const commit = await prisma.commit.create({
@@ -329,18 +418,27 @@ commitRouter.post('/create', async (req, res) => {
                 status,
                 branchId: branchId,            // makes the necessary updates in the branch model
                 commitHash,
+                architecture,                  // architecuture is required in the current version
                 params: {
                     // creates a new entry in the parameter schema containing the parameters for this commit
                     create: {
                         params: params.params, // Assumes params is an object with base64 encoded data
-                        ZKMLProof: {
-                            create: {
-                                verifierKey,
-                                circuitSettingsSer,
-                                proofSer,
-                                srsSer
+                        ...(
+                            params.zkmlProof.verifierKey &&
+                            params.zkmlProof.circuitSettingsSer &&
+                            params.zkmlProof.proofSer &&
+                            params.zkmlProof.srsSer &&
+                            {
+                                ZKMLProof: {
+                                    create: {
+                                        verifierKey: params.zkmlProof.verifierKey,
+                                        circuitSettingsSer: params.zkmlProof.circuitSettingsSer,
+                                        proofSer: params.zkmlProof.proofSer,
+                                        srsSer: params.zkmlProof.srsSer
+                                    }
+                                }
                             }
-                        }
+                        )
                     },
                 },
                 committerId: committer.id      // makes the necessary updates in the user schema
@@ -358,7 +456,7 @@ commitRouter.post('/create', async (req, res) => {
                 data: { updatedAt: new Date() }
             })
         ]);
-        res.status(201).json({ data: commit });
+        res.status(201).json({ data: commit, warnings });
     } catch (error) {
         console.error('Error creating commit:', error);
         res.status(500).send({ error: { message: 'Internal Server Error' } });
